+++ testdata/diff/online/add_constraint/plan.sql
@@ -1,4 +1,4 @@
 ALTER TABLE orders
-ADD CONSTRAINT check_amount_positive CHECK (amount > 0::numeric) NOT VALID;
+ADD CONSTRAINT check_amount_positive CHECK (amount > 0) NOT VALID;
 
 ALTER TABLE orders VALIDATE CONSTRAINT check_amount_positive;
